\name{apc.identify.mixed}
\alias{apc.identify.mixed}
\title{Identification of time effects for mixed frequency data}
\description{Generalizes \code{\link{apc.identify}} to mixed frequency data.}
\usage{apc.identify.mixed(apc.fit.model)}
\arguments{
  \item{apc.fit.model}{List. See \code{\link{apc.fit.model}} for a description of the format.}
}
\references{ Nielsen, B. (2022) Age-period-cohort analysis of mixed frequency data . \emph{Download}: \href{https://www.nuffield.ox.ac.uk/economics/Papers/2022/2022-W02apc_mixed.pdf}{Nuffield Discussion Paper 2022-W02}.
}
\author{Bent Nielsen <bent.nielsen@nuffield.ox.ac.uk> 3 Jul 2025}


