\name{apc.test.normal.residuals}
\alias{apc.test.normal.residuals}
\title{Test normality of residuals}
\description{Function that tests normality of vector of residuals.}
\usage{apc.test.normal.residuals(list_or_vector,remove.zeros=FALSE,tolerance=1e-10)}
\arguments{
    \item{list_or_vector}{EITHER list with object residuals	OR vector of residuals.}
    \item{remove.zeros}{Optional. Logical. Default is FALSE.}
    \item{tolerance}{Optional. Default is 1e-10.}
    
}