% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAPC.R
\name{calcAPC}
\alias{calcAPC}
\title{Calculate the annual porcentual change}
\usage{
calcAPC(fdata, f_ph)
}
\arguments{
\item{fdata}{Data frame which has on its first column the subject ID and on the rest of the columns the data to calculate the APC from.}

\item{f_ph}{Data frame, output from function load_phenotypes().}
}
\value{
Data frame
}
\description{
This function calculates the annual porcentual change (APC) of a matrix which has the subject ID as its first column and on the rest of the columns the numeric data to calculate the APC from. Therefore, each row is a different scan ID and each column a different region.
}
\author{
Lucia de Hoyos <luciadeh@gmail.com>
}
