\name{data.Swiss.suicides}
\alias{data.Swiss.suicides}
\title{Swiss suicide data}
\description{Function that organises Swiss suicide data. Data are used in Nielsen (2022) to illustrate the 2-sample age-period-cohort analysis. The analysis is presented in a vignette \href{https://users.ox.ac.uk/~nuff0078/apc/vignettes/ReproducingN2025.pdf}{ReproducingN2025.pdf}, \href{https://users.ox.ac.uk/~nuff0078/apc/vignettes/ReproducingN2025.R}{ReproducingN2025.R} on \href{https://users.ox.ac.uk/~nuff0078/apc/index.htm}{vignette}. Source: Riebler, Held, Rue and Bopp (2012).}
\usage{data.Swiss.suicides()}
\author{Bent Nielsen <bent.nielsen@nuffield.ox.ac.uk> 7 Jul 2025}
\references{Nielsen, B. (2022) Two-sample age-period-cohort models with an application to Swiss suicide rates. \emph{Download}: \href{https://www.nuffield.ox.ac.uk/economics/Papers/2022/2022-W03apc_2sample.pdf}{Nuffield Discussion Paper 2022-W03}.

    Riebler, A. and Held, L. and Rue, H. and Bopp, M. (2012) Gender-specific differences and the impact of family integration on time trends in age-stratified Swiss suicide rates. \emph{Journal of the Royal Statistical Society, Series A}, 175, 479-490.}
