% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_phenotypes.R
\name{load_phenotypes}
\alias{load_phenotypes}
\title{Prepare the phenotypic data}
\usage{
load_phenotypes(file, colID = NULL, colsID = NULL, colTp = NULL, colAge = NULL)
}
\arguments{
\item{file}{a data frame with the phenotypic information of subjects.}

\item{colID}{a number defining the column where the scan ID is specified.}

\item{colsID}{a number defining the column where the subject ID is specified.}

\item{colAge}{a number defining the column where the age of the patient at the scan is specified}

\item{colTP}{a number defining the column where the timepoint is specified. All subjects must have timepoint 1.}
}
\value{
Data frame
}
\description{
This function prepares the phenotypic data of the subjects. It inputs the phenotypic information and the column numbers of scan ID (colID), subject ID (colsID), scan timepoint (colTp) and age (colAge). It reorganises the information, to output a data frame which has the information needed to calculate the APCs for each subject and its two timepoints (scan ID and age at first and second timepoint).
}
\examples{
y <- load_phenotypes(file= mtcars, colID = 1, colsID = 2, colTp = 3, colAge = 4)

}
\author{
Lucia de Hoyos <luciadeh@gmail.com>
}
