% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSC.R
\name{runSC}
\alias{runSC}
\title{Compute and compare the SC matrices}
\usage{
runSC(data1, data2)
}
\arguments{
\item{data1}{Data frame 1 with information to be converted into a SC matrix. Do not input subject ID, only metrics values.}

\item{data2}{Data frame 2 with information to be converted into a SC matrix. Do not input subject ID, only metrics values.}
}
\value{
Matrix of p-values.
}
\description{
This function creates the structural covariance (SC) matrices from the two input datasets (data1, data2) and then compares them using Z-scores. The function then outputs a matrix with p-values that indicate which differences between the two matrices are significant. To do the comparison, the function substracts data1 to data2.
}
\author{
Lucia de Hoyos <luciadeh@gmail.com>
}
