\name{apc-internal}
\alias{apc.internal.function.date.2.character}
\alias{foo2}
\alias{foo3}
\alias{foo4}
\title{Internal apc Functions}
\description{
  Internal apc functions
}
\details{
  These are not to be called by the user.
}
\author{Bent Nielsen <bent.nielsen@nuffield.ox.ac.uk> 1 Feb 2016}